IndexDeterminant added, Interpolator started, ColumnTemperature started

+++ FireResistance.Core/Entities/Constructions/AbstractClasses/Column.cs
@@ -9,9 +9,9 @@ namespace FireResistance.Core.Entities.Constructions.AbstractClasses
 {
     internal class Column : Construction
     {
-        public virtual double distanceToArmature { get; set; }
+        public virtual int distanceToArmature { get; set; }
         public virtual string fixationElement { get; set; }
-        public virtual double workLenth { get; set; }
+        public virtual int workLenth { get; set; }
 
         public override string ToString()
         {
